# speed up ActionDispatch::Journey::Route#matches?