@kaspth 2016-06-26
Hey, thanks for looking into this 👋

How often is `matches?` called in a request? I mean, it would be nice if we could benchmark if this change had an overall effect on request time. Try using the benchmarks here: https://github.com/eileencodes/integration_performance_test

cc @pixeltrix