@rafaelfranca 2016-06-26
Yeah, what is the motivation of this speedup? The code now is way harder to understand. That `!` in the `case` statement is very confusion. If the only motivation is by a micro benchmark I prefer to not apply the patch as it is.