@kaspth 2016-08-16
I'm going to close this. The code is harder to read and we haven't seen a benchmark for the whole request. Thanks for the pull request! ❤️